Changed GameManager
Changed the number of brick rows to be a set number and changed brick setup algorithm to modulo on the number of colors when creating a brick of what color. This should prevent any errors no matter the number of bricks or rows chosen with the index going out of bounds.